refactoring(패키지 이동) : 문자열계산기 패키지 이동( lotto -> calculator)
