+++ src/main/java/calculator/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package lotto;
+package calculator;
 
 import java.util.Optional;
 import java.util.regex.Matcher;
