+++ src/test/java/calculator/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package calculator;
 
 import org.junit.jupiter.api.Test;
